+++ ios/Classes/FlutterLinkedinLoginPlugin.m
@@ -1,7 +1,7 @@
 #import "FlutterLinkedinLoginPlugin.h"
 #import <linkedin-sdk/LISDK.h>
 
-NSString *const URL = @"https://api.linkedin.com/v1/people/~:(id,first-name,last-name,headline,industry,summary,picture-url)?format=json";
+NSString *const URL = @"https://api.linkedin.com/v1/people/~:(id,first-name,last-name,headline,industry,summary,picture-url,email-address,formatted-name,location,specialties,positions)?format=json";
 
 /**
  * FlutterLinkedInLoginPlugin handles login, clearing session, and getting basic user
@@ -23,9 +23,9 @@ NSString *const URL = @"https://api.linkedin.com/v1/people/~:(id,first-name,last
         //NSLog(@"%s","In method caller");
         if ([call.method isEqual: @"logIntoLinkedIn"]) {
             [self logIntoLinkedIn:result];
-        } else if ([call.method isEqual: @"logIntoLinkedInBasic"]) {
+        } else if ([call.method isEqual: @"loginBasic"]) {
             [self logIntoLinkedIn:result];
-        } else if ([call.method isEqual: @"logIntoLinkedInFull"]) {
+        } else if ([call.method isEqual: @"loginBasicWithProfile"]) {
             [self logIntoLinkedIn:result];
         } else if ([call.method isEqual: @"getProfile"]) {
             [self getProfile:result];