+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2018-09-09
+### Added
+- emailAddress, formattedName, and location to LinkedInProfile
+- toJson() for LinkedInProfile
+- methods loginBasic(), logout(), and getProfile() in preparation to remove getters
+
 ## 1.0.0 - 2018-07-22
 ### Added
 - Support for iOS